rename connections to integations
