+++ packages/eas-cli/src/commands/integrations/asc/__tests__/disconnect.test.ts
@@ -3,13 +3,17 @@ import { GraphQLError } from 'graphql/error';
 
 import { getMockOclifConfig } from '../../../../__tests__/commands/utils';
 import { ExpoGraphqlClient } from '../../../../commandUtils/context/contextUtils/createGraphqlClient';
+import { AscAppLinkMutation } from '../../../../graphql/mutations/AscAppLinkMutation';
 import { AscAppLinkQuery } from '../../../../graphql/queries/AscAppLinkQuery';
 import Log from '../../../../log';
-import ConnectionsAscStatus from '../status';
+import { toggleConfirmAsync } from '../../../../prompts';
+import IntegrationsAscDisconnect from '../disconnect';
 
 jest.mock('../../../../graphql/queries/AscAppLinkQuery');
+jest.mock('../../../../graphql/mutations/AscAppLinkMutation');
 jest.mock('../../../../log');
 jest.mock('../../../../ora');
+jest.mock('../../../../prompts');
 
 const testProjectId = 'test-project-id';
 const mockMetadataConnected = {
@@ -35,7 +39,7 @@ const mockMetadataDisconnected = {
   appStoreConnectApp: null,
 };
 
-describe(ConnectionsAscStatus, () => {
+describe(IntegrationsAscDisconnect, () => {
   const graphqlClient = {} as any as ExpoGraphqlClient;
   const mockConfig = getMockOclifConfig();
 
@@ -43,10 +47,14 @@ describe(ConnectionsAscStatus, () => {
     jest.clearAllMocks();
   });
 
-  it('displays connected status', async () => {
-    jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockResolvedValueOnce(mockMetadataConnected);
+  it('disconnects successfully with --yes', async () => {
+    jest
+      .mocked(AscAppLinkQuery.getAppMetadataAsync)
+      .mockResolvedValueOnce(mockMetadataConnected)
+      .mockResolvedValueOnce(mockMetadataDisconnected);
+    jest.mocked(AscAppLinkMutation.deleteAppStoreConnectAppAsync).mockResolvedValueOnce(undefined);
 
-    const command = new ConnectionsAscStatus([], mockConfig);
+    const command = new IntegrationsAscDisconnect(['--yes'], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -54,16 +62,27 @@ describe(ConnectionsAscStatus, () => {
     });
 
     await command.runAsync();
-    expect(AscAppLinkQuery.getAppMetadataAsync).toHaveBeenCalledWith(graphqlClient, testProjectId);
-    expect(jest.mocked(Log.log)).toHaveBeenCalled();
+
+    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).toHaveBeenCalledWith(
+      graphqlClient,
+      'asc-app-link-id'
+    );
+    expect(AscAppLinkQuery.getAppMetadataAsync).toHaveBeenNthCalledWith(
+      2,
+      graphqlClient,
+      testProjectId,
+      {
+        useCache: false,
+      }
+    );
   });
 
-  it('displays disconnected status', async () => {
+  it('no-op when already disconnected', async () => {
     jest
       .mocked(AscAppLinkQuery.getAppMetadataAsync)
       .mockResolvedValueOnce(mockMetadataDisconnected);
 
-    const command = new ConnectionsAscStatus([], mockConfig);
+    const command = new IntegrationsAscDisconnect(['--yes'], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -71,15 +90,17 @@ describe(ConnectionsAscStatus, () => {
     });
 
     await command.runAsync();
+
+    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
     expect(jest.mocked(Log.log)).toHaveBeenCalled();
   });
 
-  it('prints json output in json mode', async () => {
+  it('prints json output when already disconnected in json mode', async () => {
     jest
       .mocked(AscAppLinkQuery.getAppMetadataAsync)
       .mockResolvedValueOnce(mockMetadataDisconnected);
 
-    const command = new ConnectionsAscStatus(['--json'], mockConfig);
+    const command = new IntegrationsAscDisconnect(['--yes', '--json'], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -87,8 +108,10 @@ describe(ConnectionsAscStatus, () => {
     });
 
     await command.runAsync();
+
+    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
     expect(jest.mocked(Log.log)).toHaveBeenCalledWith(
-      expect.stringContaining('"action": "status"')
+      expect.stringContaining('"action": "disconnect"')
     );
   });
 
@@ -102,7 +125,7 @@ describe(ConnectionsAscStatus, () => {
     });
     jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockRejectedValueOnce(ascError);
 
-    const command = new ConnectionsAscStatus([], mockConfig);
+    const command = new IntegrationsAscDisconnect(['--yes'], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -110,46 +133,28 @@ describe(ConnectionsAscStatus, () => {
     });
 
     await command.runAsync();
+    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
     expect(jest.mocked(Log.warn)).toHaveBeenCalledWith(
       expect.stringContaining('revoked or is no longer valid')
     );
   });
 
-  it('handles invalid ASC API key with json output', async () => {
-    const ascError = new CombinedError({
-      graphQLErrors: [
-        new GraphQLError(
-          'App Store Connect rejected this API key with status 401. Choose a valid API key and try again.'
-        ),
-      ],
-    });
-    jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockRejectedValueOnce(ascError);
+  it('cancels disconnection when confirmation is rejected', async () => {
+    const processExitSpy = jest.spyOn(process, 'exit').mockImplementation((() => {
+      throw new Error('process.exit called');
+    }) as never);
+    jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockResolvedValueOnce(mockMetadataConnected);
+    jest.mocked(toggleConfirmAsync).mockResolvedValueOnce(false);
 
-    const command = new ConnectionsAscStatus(['--json'], mockConfig);
+    const command = new IntegrationsAscDisconnect([], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
       loggedIn: { graphqlClient },
     });
 
-    await command.runAsync();
-    expect(jest.mocked(Log.log)).toHaveBeenCalledWith(
-      expect.stringContaining('"status": "invalid"')
-    );
-  });
-
-  it('throws when fetching status fails', async () => {
-    jest
-      .mocked(AscAppLinkQuery.getAppMetadataAsync)
-      .mockRejectedValueOnce(new Error('status failed'));
-
-    const command = new ConnectionsAscStatus([], mockConfig);
-    // @ts-expect-error
-    jest.spyOn(command, 'getContextAsync').mockReturnValue({
-      projectId: testProjectId,
-      loggedIn: { graphqlClient },
-    });
-
-    await expect(command.runAsync()).rejects.toThrow('status failed');
+    await expect(command.runAsync()).rejects.toThrow('process.exit called');
+    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
+    processExitSpy.mockRestore();
   });
 });
