+++ packages/eas-cli/src/commands/integrations/asc/__tests__/status.test.ts
@@ -3,17 +3,13 @@ import { GraphQLError } from 'graphql/error';
 
 import { getMockOclifConfig } from '../../../../__tests__/commands/utils';
 import { ExpoGraphqlClient } from '../../../../commandUtils/context/contextUtils/createGraphqlClient';
-import { AscAppLinkMutation } from '../../../../graphql/mutations/AscAppLinkMutation';
 import { AscAppLinkQuery } from '../../../../graphql/queries/AscAppLinkQuery';
 import Log from '../../../../log';
-import { toggleConfirmAsync } from '../../../../prompts';
-import ConnectionsAscDisconnect from '../disconnect';
+import IntegrationsAscStatus from '../status';
 
 jest.mock('../../../../graphql/queries/AscAppLinkQuery');
-jest.mock('../../../../graphql/mutations/AscAppLinkMutation');
 jest.mock('../../../../log');
 jest.mock('../../../../ora');
-jest.mock('../../../../prompts');
 
 const testProjectId = 'test-project-id';
 const mockMetadataConnected = {
@@ -39,7 +35,7 @@ const mockMetadataDisconnected = {
   appStoreConnectApp: null,
 };
 
-describe(ConnectionsAscDisconnect, () => {
+describe(IntegrationsAscStatus, () => {
   const graphqlClient = {} as any as ExpoGraphqlClient;
   const mockConfig = getMockOclifConfig();
 
@@ -47,14 +43,10 @@ describe(ConnectionsAscDisconnect, () => {
     jest.clearAllMocks();
   });
 
-  it('disconnects successfully with --yes', async () => {
-    jest
-      .mocked(AscAppLinkQuery.getAppMetadataAsync)
-      .mockResolvedValueOnce(mockMetadataConnected)
-      .mockResolvedValueOnce(mockMetadataDisconnected);
-    jest.mocked(AscAppLinkMutation.deleteAppStoreConnectAppAsync).mockResolvedValueOnce(undefined);
+  it('displays connected status', async () => {
+    jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockResolvedValueOnce(mockMetadataConnected);
 
-    const command = new ConnectionsAscDisconnect(['--yes'], mockConfig);
+    const command = new IntegrationsAscStatus([], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -62,45 +54,16 @@ describe(ConnectionsAscDisconnect, () => {
     });
 
     await command.runAsync();
-
-    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).toHaveBeenCalledWith(
-      graphqlClient,
-      'asc-app-link-id'
-    );
-    expect(AscAppLinkQuery.getAppMetadataAsync).toHaveBeenNthCalledWith(
-      2,
-      graphqlClient,
-      testProjectId,
-      {
-        useCache: false,
-      }
-    );
-  });
-
-  it('no-op when already disconnected', async () => {
-    jest
-      .mocked(AscAppLinkQuery.getAppMetadataAsync)
-      .mockResolvedValueOnce(mockMetadataDisconnected);
-
-    const command = new ConnectionsAscDisconnect(['--yes'], mockConfig);
-    // @ts-expect-error
-    jest.spyOn(command, 'getContextAsync').mockReturnValue({
-      projectId: testProjectId,
-      loggedIn: { graphqlClient },
-    });
-
-    await command.runAsync();
-
-    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
+    expect(AscAppLinkQuery.getAppMetadataAsync).toHaveBeenCalledWith(graphqlClient, testProjectId);
     expect(jest.mocked(Log.log)).toHaveBeenCalled();
   });
 
-  it('prints json output when already disconnected in json mode', async () => {
+  it('displays disconnected status', async () => {
     jest
       .mocked(AscAppLinkQuery.getAppMetadataAsync)
       .mockResolvedValueOnce(mockMetadataDisconnected);
 
-    const command = new ConnectionsAscDisconnect(['--yes', '--json'], mockConfig);
+    const command = new IntegrationsAscStatus([], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -108,10 +71,24 @@ describe(ConnectionsAscDisconnect, () => {
     });
 
     await command.runAsync();
+    expect(jest.mocked(Log.log)).toHaveBeenCalled();
+  });
 
-    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
+  it('prints json output in json mode', async () => {
+    jest
+      .mocked(AscAppLinkQuery.getAppMetadataAsync)
+      .mockResolvedValueOnce(mockMetadataDisconnected);
+
+    const command = new IntegrationsAscStatus(['--json'], mockConfig);
+    // @ts-expect-error
+    jest.spyOn(command, 'getContextAsync').mockReturnValue({
+      projectId: testProjectId,
+      loggedIn: { graphqlClient },
+    });
+
+    await command.runAsync();
     expect(jest.mocked(Log.log)).toHaveBeenCalledWith(
-      expect.stringContaining('"action": "disconnect"')
+      expect.stringContaining('"action": "status"')
     );
   });
 
@@ -125,7 +102,7 @@ describe(ConnectionsAscDisconnect, () => {
     });
     jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockRejectedValueOnce(ascError);
 
-    const command = new ConnectionsAscDisconnect(['--yes'], mockConfig);
+    const command = new IntegrationsAscStatus([], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
@@ -133,28 +110,46 @@ describe(ConnectionsAscDisconnect, () => {
     });
 
     await command.runAsync();
-    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
     expect(jest.mocked(Log.warn)).toHaveBeenCalledWith(
       expect.stringContaining('revoked or is no longer valid')
     );
   });
 
-  it('cancels disconnection when confirmation is rejected', async () => {
-    const processExitSpy = jest.spyOn(process, 'exit').mockImplementation((() => {
-      throw new Error('process.exit called');
-    }) as never);
-    jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockResolvedValueOnce(mockMetadataConnected);
-    jest.mocked(toggleConfirmAsync).mockResolvedValueOnce(false);
+  it('handles invalid ASC API key with json output', async () => {
+    const ascError = new CombinedError({
+      graphQLErrors: [
+        new GraphQLError(
+          'App Store Connect rejected this API key with status 401. Choose a valid API key and try again.'
+        ),
+      ],
+    });
+    jest.mocked(AscAppLinkQuery.getAppMetadataAsync).mockRejectedValueOnce(ascError);
 
-    const command = new ConnectionsAscDisconnect([], mockConfig);
+    const command = new IntegrationsAscStatus(['--json'], mockConfig);
     // @ts-expect-error
     jest.spyOn(command, 'getContextAsync').mockReturnValue({
       projectId: testProjectId,
       loggedIn: { graphqlClient },
     });
 
-    await expect(command.runAsync()).rejects.toThrow('process.exit called');
-    expect(AscAppLinkMutation.deleteAppStoreConnectAppAsync).not.toHaveBeenCalled();
-    processExitSpy.mockRestore();
+    await command.runAsync();
+    expect(jest.mocked(Log.log)).toHaveBeenCalledWith(
+      expect.stringContaining('"status": "invalid"')
+    );
+  });
+
+  it('throws when fetching status fails', async () => {
+    jest
+      .mocked(AscAppLinkQuery.getAppMetadataAsync)
+      .mockRejectedValueOnce(new Error('status failed'));
+
+    const command = new IntegrationsAscStatus([], mockConfig);
+    // @ts-expect-error
+    jest.spyOn(command, 'getContextAsync').mockReturnValue({
+      projectId: testProjectId,
+      loggedIn: { graphqlClient },
+    });
+
+    await expect(command.runAsync()).rejects.toThrow('status failed');
   });
 });
